requirements.txt
numpy==1.19.3
astc_decomp
biplist
bs4
pyliblzfse
packaging==20.1
pathlib2==2.3.5
pillow
pyinstaller
PySimpleGUI==4.16.0
blackboxprotobuf
nska-deserialize>=1.3.1
biplist
simplekml
pandas

